move function to geometryHelpers
