+++ src/DashboardUI/src/utilities/geometryHelpers.ts
@@ -1,5 +1,7 @@
 import { circle } from '@turf/circle';
 import { distance } from '@turf/distance';
+import { featureCollection } from '@turf/helpers';
+import intersect from '@turf/intersect';
 import { Feature, FeatureCollection, GeoJsonProperties, Geometry, Polygon, Position } from 'geojson';
 import mapboxgl from 'mapbox-gl';
 
@@ -35,4 +37,21 @@ export const generateCircleWithRadiusFromCenterPointToEdgePoint = (
     units: 'kilometers',
   });
   return circle(circleCenterPoint, distanceFromCenterToEdgeInKm, { steps: 100 });
+};
+
+export const doPolygonsIntersect = (polygons: Feature<Geometry, GeoJsonProperties>[]): boolean => {
+  for (let i = 0; i < polygons.length; i++) {
+    for (let j = i + 1; j < polygons.length; j++) {
+      const p1 = polygons[i] as Feature<Polygon, GeoJsonProperties>;
+      const p2 = polygons[j] as Feature<Polygon, GeoJsonProperties>;
+
+      const intersection = intersect(featureCollection([p1, p2]));
+
+      if (intersection) {
+        return true;
+      }
+    }
+  }
+
+  return false;
 };
